pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "plexus-agent"
version = "0.5.0"
description = "Send sensor data to Plexus in one line of code"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.8"
authors = [
    { name = "Plexus", email = "hello@plexus.dev" }
]
keywords = ["sensors", "telemetry", "iot", "hardware", "observability", "robotics"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: System :: Hardware",
]
dependencies = [
    "requests>=2.28.0",
    "click>=8.0.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
mqtt = ["paho-mqtt>=1.6.0"]
ros = [
    "rosbags>=0.9.0",           # ROS1/ROS2 bag reading
    "mcap>=1.0.0",               # MCAP format support
]
camera = [
    "opencv-python>=4.8.0",      # USB webcam support
    "numpy>=1.20.0",
]
picamera = [
    "picamera2>=0.3.12",         # Raspberry Pi Camera Module
]
sensors = [
    "smbus2>=0.4.0",             # I2C communication for sensors
]
can = [
    "python-can>=4.0.0",         # CAN bus interface
    "cantools>=39.0.0",          # DBC file parsing
]
serial = [
    "pyserial>=3.5",             # Serial port communication
]
tui = [
    "rich>=13.0.0",              # Rich terminal output and live dashboard
]
system = [
    "psutil>=5.9.0",             # System resource monitoring
]
all = [
    "paho-mqtt>=1.6.0",
    "rosbags>=0.9.0",
    "mcap>=1.0.0",
    "opencv-python>=4.8.0",
    "numpy>=1.20.0",
    "smbus2>=0.4.0",
    "python-can>=4.0.0",
    "cantools>=39.0.0",
    "pyserial>=3.5",
    "rich>=13.0.0",
    "psutil>=5.9.0",
]
dev = ["pytest", "pytest-cov", "ruff"]

[project.scripts]
plexus = "plexus.cli:main"

[project.urls]
Homepage = "https://plexus.dev"
Documentation = "https://docs.plexus.dev"
Repository = "https://github.com/plexus-oss/agent"
Issues = "https://github.com/plexus-oss/agent/issues"

[tool.hatch.build.targets.wheel]
packages = ["plexus"]

[tool.ruff]
line-length = 100
target-version = "py38"
